+++ internal/FViz/Rendering/FVizActorPrivate.h
@@ -2,6 +2,8 @@
 #define FVIZ_INTERNAL_RENDERING_ACTOR_PRIVATE_H
 
 #include <FViz/Core/FVizObjectPrivate.h>
+#include <FViz/Math/FVizQuat.h>
+#include <FViz/Math/FVizVec3.h>
 #include <FViz/Rendering/FVizActor.h>
 
 struct FVizActor
@@ -11,6 +13,9 @@ struct FVizActor
     float color[3];
     FVizBool visible;
     FVizBool wireframe;
+    FVizVec3 position;
+    FVizQuat orientation;
+    FVizVec3 scale;
 };
 
 #endif /* FVIZ_INTERNAL_RENDERING_ACTOR_PRIVATE_H */